src/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.128.0",
    "google-cloud-aiplatform>=1.135.0",
    "google-cloud-bigquery>=3.40.0",
    "pydantic-settings>=2.12.0",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.14.14",
    "ty>=0.0.14",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ty.src]
include = ["**/*.py"]

[tool.ruff]
line-length = 120
